+++ src/main/java/com/bangil/toy_sns/service/board/BoardService.java
@@ -4,6 +4,6 @@ import com.bangil.toy_sns.domain.Board;
 import com.bangil.toy_sns.domain.BoardFileAttach;
 
 public interface BoardService {
-    public Long addBoard(Board board);
+    public int addBoard(Board board);
     public void addBoardFileAttach(BoardFileAttach boardFileAttach);
 }
